Move the partition builder into its own file
